requirements.txt
Flask>=2.0.0
gunicorn>=20.0.0
yt-dlp>=2023.0.0
openai>=1.0.0
requests>=2.0.0
webvtt-py>=0.4.6
